refactor: rename subject domain to sections
