+++ docs/changelog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Subject domain module
+- Sections module
 - ClangFormat for code refactoring
 - Base structure of the Metasystem
 - sc-component-manager paths for components
