+++ platform-dependent-components/problem-solver/cxx/CMakeLists.txt
@@ -1,2 +1,2 @@
 add_subdirectory(identifiers-module)
-add_subdirectory(subject-domain-module)
+add_subdirectory(sections-module)
