+++ platform-dependent-components/problem-solver/cxx/sections-module/constants/sections_aliases.cpp
@@ -1,0 +1,16 @@
+/*
+ * This source file is part of an OSTIS project. For the latest info, see http://ostis.net
+ * Distributed under the MIT License
+ * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
+ */
+
+#include "sections_aliases.hpp"
+
+std::string const sections_aliases::DECOMPOSITION_TUPLE = "_decomposition_tuple";
+std::string const sections_aliases::EDGE_TO_DELETE = "_edge_to_delete";
+std::string const sections_aliases::PARENT_SECTION = "_parent_section";
+std::string const sections_aliases::BASE_SEQUENCE_EDGE = "_base_sequence_edge";
+std::string const sections_aliases::SECTION_EDGE = "_section_edge";
+std::string const sections_aliases::SECTION_NODE = "_section";
+std::string const sections_aliases::PREVIOUS_SECTION_ARC = "_previous_section_arc";
+std::string const sections_aliases::NEXT_SECTION_ARC = "_next_section_arc";
