+++ platform-dependent-components/problem-solver/cxx/sections-module/constants/sections_aliases.hpp
@@ -1,0 +1,22 @@
+/*
+ * This source file is part of an OSTIS project. For the latest info, see http://ostis.net
+ * Distributed under the MIT License
+ * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
+ */
+
+#pragma once
+
+#include <string>
+
+class sections_aliases
+{
+public:
+  static std::string const DECOMPOSITION_TUPLE;
+  static std::string const EDGE_TO_DELETE;
+  static std::string const PARENT_SECTION;
+  static std::string const BASE_SEQUENCE_EDGE;
+  static std::string const SECTION_EDGE;
+  static std::string const SECTION_NODE;
+  static std::string const PREVIOUS_SECTION_ARC;
+  static std::string const NEXT_SECTION_ARC;
+};
