+++ platform-dependent-components/problem-solver/cxx/sections-module/subject_domain_module.hpp
@@ -1,0 +1,21 @@
+/*
+ * This source file is part of an OSTIS project. For the latest info, see http://ostis.net
+ * Distributed under the MIT License
+ * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
+ */
+
+#pragma once
+
+#include "sc-memory/sc_module.hpp"
+
+#include "generated/subject_domain_module.generated.hpp"
+
+class SectionsModule : public ScModule
+{
+  SC_CLASS(LoadOrder(100))
+  SC_GENERATED_BODY()
+
+  virtual sc_result InitializeImpl() override;
+
+  virtual sc_result ShutdownImpl() override;
+};
