+++ platform-dependent-components/problem-solver/cxx/sections-module/test/testStructures/test_add_section.scs
