+++ platform-dependent-components/problem-solver/cxx/sections-module/test/testStructures/test_remove_section.scs
