+++ platform-dependent-components/problem-solver/cxx/sections-module/test/testStructures/test_successful_decomposition.scs
