--- platform-dependent-components/problem-solver/cxx/subject-domain-module/utils/subject_domain_utils.hpp
@@ -1,32 +1,0 @@
-/*
- * This source file is part of an OSTIS project. For the latest info, see http://ostis.net
- * Distributed under the MIT License
- * (See accompanying file COPYING.MIT or copy at http://opensource.org/licenses/MIT)
- */
-
-#pragma once
-
-#include "sc-memory/sc_memory.hpp"
-
-#include "sc-agents-common/utils/IteratorUtils.hpp"
-#include "sc-agents-common/keynodes/coreKeynodes.hpp"
-
-#include "keynodes/subject_domain_keynodes.hpp"
-
-#include <string>
-
-namespace subjectDomainModule
-{
-
-class subject_domain_utils
-{
-public:
-  static ScAddr FindSectionByName(ScMemoryContext *, std::string const &);
-
-  static ScAddr GetLastSubSection(ScMemoryContext * context, ScAddr const & tuple);
-
-  // Get decomposition tuple if it exists or generate it
-  static ScAddr GetSectionDecompositionTuple(ScMemoryContext *, ScAddr const &);
-};
-
-}  // namespace subjectDomainModule
